test: rename/split daemon-channel tests to satisfy PS-204/PS-205 mirror rules
The audit-project gate (PS-204 orphan-test-file, PS-205 test-name-prefix
-mismatch) requires each test file to mirror exactly one src module:

- tests/.../_lifecycle/test_instances.py → test__instances.py
  (private _instances.py needs the double-underscore test name)
- tests/.../_runners/test_session_channels.py (orphan) split into:
  - test__session_cli.py            → mirrors _session_cli.py (--channels parse)
  - test_claude_session_channels.py → mirrors claude_session.py (run() forwarding)
  - test__session_conversation.py   → mirrors _session_conversation.py (extra-packing)

No assertion changes; pure relocation. All 14 cases still pass.
